packages/jaeger_input_otel/_dev/deploy/docker/requirements.txt
opentelemetry-api>=1.3.0
opentelemetry-sdk>=1.11.0
opentelemetry-exporter-jaeger-thrift>=1.11.0
deprecated>=1.2.0
thrift>=0.10.0
